Numerous changes (quite significant) : - Spritesheets are now vertical instead of horizontal. The reason for this is because on some platforms, it is easier and faster to cut up a portion of an image if it is vertically aligned than if it is horizontally aligned. (due to how memory works)
- Images/Sprites now also require an option for relative Width/Height.
It accepts values from 0 to 100, which is then divided by 100.
1.0 is the size of the resolution's width or height, 0.0 is 0.

This was done as to make it easier to support the game accross multiple
resolutions and aspect ratios. This also means that we don't need
to provide multiple assets depending on the resolution and we just
let the hardware do the work, which is much easier.

- The game itself can now interpret the array correctly and process them.
In addition, 2 sprites on screen are supported and they are displayed properly.
The basics are mostly done, aside from the option to choose your path.

- There were some internal changes too, so we are now using a structure for
retaining internal info about the image's width, texture etc...
This is to allow the engine to be a bit more flexible and have separate
image IDs for sprites and backgrounds.

- There is now a function to draw a static background.
This was again made to avoid issues with images and much simplier.
The background will be scaled up according to the screen's resolution.

The function to draw images is still retained but for now it is not used.

- There are now Macros for Sprites and Background.
They all redirect to the same function which will then internally redirect
to the structure's background or sprite.